+++ tools/build_audio.py
@@ -1,11 +1,18 @@
 import os
 import os.path
 import shutil
+import json
+import helpers
 
 print("\n")
 print("--------------------------------------------------------------------------------------------------------------")
 print("pmtech audio compression and conversion ----------------------------------------------------------------------")
 print("--------------------------------------------------------------------------------------------------------------")
+
+config = open("build_config.json")
+build_config = json.loads(config.read())
+
+pmtech_dir = helpers.correct_path(build_config["pmtech_dir"])
 
 platform_name = "win32"
 if os.name == "posix":
@@ -25,11 +32,12 @@ if platform_name == "osx":
 
 # copy fmod dll / dylib
 print("copying dynamic library to binary dir")
-src_file = os.path.join(os.getcwd(), "..", "pen", "third_party", "fmod", "lib", platform_name, dll)
+src_file = os.path.join(pmtech_dir, "pen", "third_party", "fmod", "lib", platform_name, dll)
 shutil.copy(src_file, bin_dir)
 
 # copy audio files
-print("copying audio to data dir")
-for f in os.listdir(audio_dir):
-    src_file = os.path.join(audio_dir,f)
-    shutil.copy(src_file, build_dir)
+if os.path.exists(audio_dir):
+    print("copying audio to data dir")
+    for f in os.listdir(audio_dir):
+        src_file = os.path.join(audio_dir,f)
+        shutil.copy(src_file, build_dir)
